pyproject.toml
[project]
name = "opticaldcn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
]
